This commit is ugly as hell and should only be used by a developer who runs it in debug mode! This was only meant for a one time import of all of my contacts on a new device (so I didn't need to sync) Added code to restore all fields and not just some. SUPER ugly mix of old and new contacts API (needs 2.0 api)! You can also specify to which account type and which account the contacts are restored.
Tested this with a fake exchange account. I don't sync my contacts at all! So I don't know if this will work with an actual exchange account or when you actually try to sync the added contacts.

I normally wouldn't upload code that is in such a bad shape, but I though it might be useful for others...